ceph.spec.in: remove build directory at end of %install
By the time we get to the end of the %install section, all the
built binaries have been installed in the build root, so we can
delete the build directory from the source tree.  This frees up
about 17GB of disk space on build hosts, which is helpful in
case other processes later in the RPM build need more disk space.

Fixes: https://tracker.ceph.com/issues/55079
Signed-off-by: Tim Serong <tserong@suse.com>