mgr/dashboard: Imrove error message of '/api/grafana/validation' API endpoint
In case the validation of the Grafana URL fails, e.g. because of an invalid SSL certificate, a useless and not helping default error message is displayed in the UI.
This PR will re-raise the exception as a DashboardException which includes the detailed description of what happened. This will help to identify SSL cert issues much easier for example.

Fixes: https://tracker.ceph.com/issues/55133

Signed-off-by: Volker Theile <vtheile@suse.com>
(cherry picked from commit 83524ed1bfb7c8cef307d7c21c3452cbe7a16c4b)