rgw: allow S3 delete-marker behavior to be restored via config
According to S3 documentation a removal of an already-removed object
from a versioned bucket creates an additional delete-marker. This has
the potential to create a huge number of delete-markers for an
object. An earlier commit inhibited this behavior and did not add
additional delete-markers. This commit adds a configuratrion option --
rgw_enable_s3_consecutive_delete_markers -- that allows the standard
S3 behavior to be restored should a client rely upon it.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>