cephfs-shell: allow creating hard and soft links
Add ln command to cephfs-shell for enabling creation of hard and soft
links. It allows creating hard links to regular files and soft links to
regular files as well as directories. The behaviour of this cephfs-shell
command is kept as close as possible to the ln command from GNU
coreutils.

Test and docs too have been updated for the same.

Fixes: https://tracker.ceph.com/issues/48911

Signed-off-by: Dhairya Parmar <dparmar@redhat.com>