mon/Elector: notify_rank_removed correctly nuke highest ranked MON
Added a case where we are removing the highest rank monitor
in `notify_rank_removed`, the old version did not deal with this
since it would only go into the loop when rank_removed < paxos_size().
Therefore, we added an else case for when rank_removed == paxos_size(),
we erase the rank from both `live_pinging` and `dead_pinging` set.

Fixes: https://tracker.ceph.com/issues/55435

Signed-off-by: Kamoltat <ksirivad@redhat.com>