cmake/modules/BuildFIO: specify the full path to alloc libraries
before this change, we assume that "ALLOCATOR" is the library name
which can be found by linker with "ld -l <namespec>". but ideally,
cmake can find a library in a more sophiscated way used by
its "find_library()" implementation. so, in this change, instead of
relying on the default paths looked up by "ld", use the path
found by cmake.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>