crimson/os/alienstore: Change policy of creating ceph_context
AlienStore used to always create ceph_context.
It was perfectly fine for crimson builds that needed classic environment
only when AlienStore instantialized BlueStore.
This is wrong when AlienStore is used as a component of classic.

Now create ceph_context only if it was null.
This allows to adapt for both cases:
- alienstore created for crimson-osd
- alienstore created for xstore (classic)

Signed-off-by: Adam Kupczyk <akupczyk@redhat.com>