mgr/dashboard: avoid overriding the configuration in package.json
the warning in npm run build is `Option "configuration" was already specified with value "en-US". The
new value "production" will override it.`

This causes the dist for production env to be created in dist/ whereas
it should be dist/en-US (or whatever npm_config_locale. That prevented
from dashboard to load and cause the cephadm e2e checks to fail.

https://jenkins.ceph.com/job/ceph-dashboard-cephadm-e2e-nightly-master/185/consoleFull#57532618373a8703b-5adb-41c5-84a0-8cf4e065ba3d

Related to https://github.com/ceph/ceph/pull/46143

Signed-off-by: Nizamudeen A <nia@redhat.com>