rgw: add functionality of SetAttribute and AddEvent method in postRequest context
opentelemetry supports mainly string, int64, double and boolean for values of trace's Attributes,
so we should validate those types and static cast to the proper type, which is different than Lua types

SetAttribute Closure will be returned to lua only if the request's trace is real and not noop span or even null like what happens in preRequest Context

AddEvent method comes to give us the ability to record an event. the event can be a single string that represents the event, or event name and key-value pairs.

Signed-off-by: Omri Zeneva <ozeneva@redhat.com>