rgw: remove potentially conficting definition of dout_subsys
dout_subsys should strictly be defined in .cc files. There was a
definition of dout_subsys in rgw_log.h, which is included in a variety
of .cc files due to transitive includes. When a subsystem wants a
conflicting definition, a compiler warning is then generated due to
the redefinition of dout_subsys.

The definition is removed from rgw_log.h and follow-on fixes are made.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>
(cherry picked from commit 8efea3e894e7db82a6f65cd997bb832e1e91a54a)