osd: optimize pg peering latency when add new osd that need backfill
set last_backfill to MIN when creating pg

This happens when the newly created(caused by adding or deleting osd)
pg has its pglog (head, tail) that is empty pglog(0,0),
but it still continuous with the authoritative log(0, 3000),
In this case, we use backfill instead of recovery.

If the authoritative log (6000,9000), that is, tail > 3000,
then the newly created pg (0,0) will naturally go directly to the backfill path,
because there is no intersection between the two.

If the osd is offline for a short time,
the pg on it will still be continuous with the authoritative log.
In this case, the recovery path will still be taken, not backfill.

There are 2 benefits
1. Backfill has lower latency than recovery in processing pglog during peering
2. When choose_acting, the backfill osd will not be considered,
   when osd is continuously added/deleted, the number of acting changes can be reduced.

Fix : https://tracker.ceph.com/issues/52884

Signed-off-by: Jianwei Zhang <jianwei1216@qq.com>