RGW - Zipper - Virtual base classes
In order to support filters, make the base classes in Zipper pure
virtual.  Then, derive from them a set of Store base classes, that
implement common code used by stores, and a set of Filter base classes
that automatically pass through to the next layer.  Modify the stores to
derive from the Store base classes.

This implements the pure virtual base classes, the Store base classes,
and a framework of the Filter base classes.

Signed-off-by: Daniel Gryniewicz <dang@redhat.com>