doc/dev: refine basic-workflow.rst integ. test
This is one in a set of PRs meant to keep the Basic
Workflow in the Developer guide current. It refines
the English in the "Integration Tests AKA ceph-qa-suite"
section of "Basic Workflow".

Several other small updates like this are expected. I
intend to avoid refining half of the page in one commit,
as I did last month when I refined the first half of the
basic workflow.

Co-authored-by: Anthony D'Atri <anthony.datri@gmail.com>

Signed-off-by: Zac Dover <zac.dover@gmail.com>