doc/dev: move option -R to a different section of doc
It's incorrect to pass option "-R fail" to the teuthology-suite command
meant for triggering tests for first time. teuthology-suite command will
fail if "-R" is passed without "-r". Therefore, move this option and its
description from the section meant for triggering tests for first time
to the section dedicated to re-running of tests.

Signed-off-by: Rishabh Dave <ridave@redhat.com>