doc/start: remove "further" (unnecessary)
The word "further" appears in a sentence, and it was probably
helpful to the person who wrote the sentence as they were
thinking about what they were trying to communicate about the
CRUSH algorithm. It's just confusing to readers, though. I tested
it on several readers ignorant of Ceph, and they told me so. So:
goodbye, "further".

Signed-off-by: Zac Dover <zac.dover@gmail.com>