mgr/cephadm: expose scheduled daemon actions and daemon action history
Previously scheduled daemon actions were not visible at all
to the user. No reason we can't show them. Some daemon actions
were recorded as events for daemons but it could be annoying
to have to go check inidividual daemons for an event. This allows
a centralized place to look for recent history and the events still
exists to get a more detailed view of what was done and
try to debug failures

Partially fixes: https://tracker.ceph.com/issues/53539

Signed-off-by: Adam King <adking@redhat.com>