mds: switch to use projected inode instead
When the client send a sync request to MDS even the client has the
Xx caps, for example, the MDS will revoke the Xx caps back and journal
new changes.

If the client try to create a sub-directory the MDS was always using
the non-projected inode of parent to inherit the gid/uid/mode metadatas,
if the journal log still haven't been flushed and applied yet, it will
always get the old metadatas.

Fixes: https://tracker.ceph.com/issues/56010
Signed-off-by: Xiubo Li <xiubli@redhat.com>