rgw_sal_motr: [CORTX-32310] fix obj size on put
The regression was introduced at commit 0ed9f66eaeb, after
which the object size recorded in its metadata on put-object
operation is always aligned to the unit size. So if the real
object size is not aligned to its unit size, the recorded
size will be wrong. (So the amount of data on get-object
operation also will be wrong.)

Solution: fix the code at MotrAtomicWriter::write() to
record the real object size in its metadata record.

Signed-off-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>