rgw: When the key_id used by the encrypted file is unavailable, using head_obj to view the details of the file will also report an error
The reason for this phenomenon is that when the head_obj interface is used, it will verify whether the key_id is available in barbican. If it is not available, the interface will return an error. A more reasonable way is that the head_obj interface returns the key_id and encryption method, but does not verify whether the key_id is available.

Fixes:https://tracker.ceph.com/issues/56080

Signed-off-by: wangyingbin <ybwang0211@163.com>