src/mon/Paxos: Provide paxos with mon shutdown info
**Problem:**

mon.a
mon.b
mon.c
mon.d
mon.e

ceph -a stop mon.d
ceph mon remove d

.
.

mon.d is down but actually did not get removed and monmap still did not get updated.

We shut down mon.d, this blocks mon.a Paxos:: begin (assuming Paxos is updating),
since mon.a (leader) sends out begin message (MMonPaxos::OP_BEGIN) to all peer
monitors including mon.d since mon.get_quorum() still is not updated and still contain mon.d.
Paxos will not proceed to the commit() phase since we did not get a reply
message(MMonPaxos::OP_ACCEPT) from mon.d. The lease of one of the monitors will eventually expire
and will call for an election, now, if the monmap proposal of ("mon remove", "name": "d")
comes in before the election happens, it will get queued to pending_finisher and eventually
be discarded once the election has started. Resulting in the remove command not taking effect.

However, if the election finishes before the monmap proposal, then we will be fine
because mon.get_quorum() will get updated and we will not be sent
(MMonPaxos::OP_BEGIN) to mon.d, hence will not be blocked and will continue to the commit phase,
therefore, this problem is non-deterministic.

**Solution:**

The goal is to allow users to successfully shutdown
and remove a monitor without Paxos
rejecting the remove proposal because it needs **all** monitors
in the quorum to accept when we know there is a monitor that was purposely shut down.
We provide Paxos with the additional information that the
monitor in the quorum might not be relevant if it is already
shut down. Once monmap is changed,  and Paxos is restarted,
the information about shut down monitor will be cleared because
we will start a new election and the quorum will be up to date with
the shutdown monitor being outside the quorum.

Fixes: https://tracker.ceph.com/issues/55695

Signed-off-by: Kamoltat <ksirivad@redhat.com>